requirements.txt
tensorflow-gpu==1.10.0
keras==2.1.3
pandas
sklearn
keras_radam
h5py
tqdm